# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Radiance Dentistry API",
        "description": "API documentation for Radiance Dentistry Dashboard",
        "contact": {
            "email": "roninaruka@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://my.radiancedentistryclinic.com/",
            "description": "Live Server"
        },
        {
            "url": "http://localhost:8080",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/slots": {
            "get": {
                "summary": "Get available slots for a date",
                "description": "Returns a list of available time slots",
                "operationId": "getAvailableSlots",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date in YYYY-MM-DD format",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "date": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "slots": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/appointments": {
            "post": {
                "summary": "Create a new appointment",
                "description": "Creates an appointment and links/creates a patient record",
                "operationId": "createAppointment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date",
                                    "time",
                                    "name",
                                    "email",
                                    "phone",
                                    "captcha_token",
                                    "captcha_answer"
                                ],
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-10"
                                    },
                                    "time": {
                                        "type": "string",
                                        "example": "10:00"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "Checkup"
                                    },
                                    "doctor_id": {
                                        "description": "Optional doctor ID",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "captcha_token": {
                                        "type": "string",
                                        "example": "uuid-token"
                                    },
                                    "captcha_answer": {
                                        "type": "string",
                                        "example": "ABCD12"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Appointment created successfully"
                    }
                }
            }
        },
...
        "/api/faqs": {
            "get": {
                "summary": "Get list of active FAQs",
                "description": "Returns a list of active FAQs sorted by sort order. Optional tag filter.",
                "operationId": "getFaqs",
                "parameters": [
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Filter FAQs by a specific tag",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "question": {
                                                "type": "string"
                                            },
                                            "answer": {
                                                "type": "string"
                                            },
                                            "tag": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "sort_order": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}